# accounts-service/src/main/java/com/banking/accounts/model/Account.java
package com.banking.accounts.model;

import jakarta.persistence.*;
import lombok.*;

@Entity
@Getter
@Setter
@ToString
@AllArgsConstructor
@NoArgsConstructor
public class Account extends  BaseEntity {

    @Column(name="customer_id")
    private Long customerId;

    @Column(name="account_number", updatable = false)
    @Id
    private Long accountNumber;

    @Column(name="account_type")
    private String accountType;

    @Column(name="branch_address")
    private String branchAddress;

}