# accounts-service/src/main/resources/static/api-spec.yml
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:9000",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/account/update": {
      "put": {
        "tags": [
          "account-controller"
        ],
        "summary": "Update Account Details REST API",
        "description": "REST API to update Customer &  Account details based on a account number",
        "operationId": "updateAccountDetails",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "HTTP Status OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "417": {
            "description": "Expectation Failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "500": {
            "description": "HTTP Status Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/account/create": {
      "post": {
        "tags": [
          "account-controller"
        ],
        "summary": "Create Account REST API",
        "description": "REST API to create new Customer &  Account",
        "operationId": "createAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "HTTP Status CREATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "500": {
            "description": "HTTP Status Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/account/fetch": {
      "get": {
        "tags": [
          "account-controller"
        ],
        "summary": "Fetch Account Details REST API",
        "description": "REST API to fetch Customer &  Account details based on a mobile number",
        "operationId": "fetchAccountDetails",
        "parameters": [
          {
            "name": "mobileNumber",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "(^$|[0-9]{10})"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HTTP Status OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "500": {
            "description": "HTTP Status Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/account/delete": {
      "delete": {
        "tags": [
          "account-controller"
        ],
        "summary": "Delete Account & Customer Details REST API",
        "description": "REST API to delete Customer &  Account details based on a mobile number",
        "operationId": "deleteAccountDetails",
        "parameters": [
          {
            "name": "mobileNumber",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "(^$|[0-9]{10})"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HTTP Status OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "417": {
            "description": "Expectation Failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "500": {
            "description": "HTTP Status Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Account": {
        "type": "object",
        "description": "Schema to hold Account information",
        "properties": {
          "accountNumber": {
            "type": "integer",
            "format": "int64",
            "description": "Account Number of Bank account",
            "example": 3454433243
          },
          "accountType": {
            "type": "string",
            "description": "Account type of Bank account",
            "example": "Savings",
            "minLength": 1
          },
          "branchAddress": {
            "type": "string",
            "description": "Bank branch address",
            "example": 123,
            "minLength": 1
          }
        },
        "required": [
          "accountNumber",
          "accountType",
          "branchAddress"
        ]
      },
      "Customer": {
        "type": "object",
        "description": "Schema to hold Customer and Account information",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the customer",
            "example": "Max Verstappen",
            "maxLength": 30,
            "minLength": 5
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the customer",
            "example": "abcd@efgh.com",
            "minLength": 1
          },
          "mobileNumber": {
            "type": "string",
            "description": "Mobile Number of the customer",
            "example": 9345432123,
            "pattern": "(^$|[0-9]{10})"
          },
          "accountDto": {
            "$ref": "#/components/schemas/Account",
            "description": "Account details of the Customer"
          }
        },
        "required": [
          "email",
          "name"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Schema to hold error response information",
        "properties": {
          "apiPath": {
            "type": "string",
            "description": "API path invoked by client"
          },
          "errorCode": {
            "type": "string",
            "description": "Error code representing the error happened",
            "enum": [
              "100 CONTINUE",
              "101 SWITCHING_PROTOCOLS",
              "102 PROCESSING",
              "103 EARLY_HINTS",
              "200 OK",
              "201 CREATED",
              "202 ACCEPTED",
              "203 NON_AUTHORITATIVE_INFORMATION",
              "204 NO_CONTENT",
              "205 RESET_CONTENT",
              "206 PARTIAL_CONTENT",
              "207 MULTI_STATUS",
              "208 ALREADY_REPORTED",
              "226 IM_USED",
              "300 MULTIPLE_CHOICES",
              "301 MOVED_PERMANENTLY",
              "302 FOUND",
              "303 SEE_OTHER",
              "304 NOT_MODIFIED",
              "307 TEMPORARY_REDIRECT",
              "308 PERMANENT_REDIRECT",
              "400 BAD_REQUEST",
              "401 UNAUTHORIZED",
              "402 PAYMENT_REQUIRED",
              "403 FORBIDDEN",
              "404 NOT_FOUND",
              "405 METHOD_NOT_ALLOWED",
              "406 NOT_ACCEPTABLE",
              "407 PROXY_AUTHENTICATION_REQUIRED",
              "408 REQUEST_TIMEOUT",
              "409 CONFLICT",
              "410 GONE",
              "411 LENGTH_REQUIRED",
              "412 PRECONDITION_FAILED",
              "413 CONTENT_TOO_LARGE",
              "413 PAYLOAD_TOO_LARGE",
              "414 URI_TOO_LONG",
              "415 UNSUPPORTED_MEDIA_TYPE",
              "416 REQUESTED_RANGE_NOT_SATISFIABLE",
              "417 EXPECTATION_FAILED",
              "418 I_AM_A_TEAPOT",
              "421 MISDIRECTED_REQUEST",
              "422 UNPROCESSABLE_CONTENT",
              "422 UNPROCESSABLE_ENTITY",
              "423 LOCKED",
              "424 FAILED_DEPENDENCY",
              "425 TOO_EARLY",
              "426 UPGRADE_REQUIRED",
              "428 PRECONDITION_REQUIRED",
              "429 TOO_MANY_REQUESTS",
              "431 REQUEST_HEADER_FIELDS_TOO_LARGE",
              "451 UNAVAILABLE_FOR_LEGAL_REASONS",
              "500 INTERNAL_SERVER_ERROR",
              "501 NOT_IMPLEMENTED",
              "502 BAD_GATEWAY",
              "503 SERVICE_UNAVAILABLE",
              "504 GATEWAY_TIMEOUT",
              "505 HTTP_VERSION_NOT_SUPPORTED",
              "506 VARIANT_ALSO_NEGOTIATES",
              "507 INSUFFICIENT_STORAGE",
              "508 LOOP_DETECTED",
              "509 BANDWIDTH_LIMIT_EXCEEDED",
              "510 NOT_EXTENDED",
              "511 NETWORK_AUTHENTICATION_REQUIRED"
            ]
          },
          "errorMessage": {
            "type": "string",
            "description": "Error message representing the error happened"
          },
          "errorTime": {
            "type": "string",
            "format": "date-time",
            "description": "Time representing when the error happened"
          }
        }
      },
      "Response": {
        "type": "object",
        "description": "Schema to hold successful response information",
        "properties": {
          "statusCode": {
            "type": "string",
            "description": "Status code in the response"
          },
          "statusMsg": {
            "type": "string",
            "description": "Status message in the response"
          }
        }
      }
    }
  }
}